+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * dfd0853 - 2017-04-26: testing new restart code 
+- * cc67724 - 2017-04-26: testing new restart code 
+- * 71d7bdb - 2017-04-26: testing new restart code 
 - * 256b107 - 2017-04-26: testing 
 - * 748bef3 - 2017-04-26: Fixed issues with restart JS code 
 - * b1eeac6 - 2017-04-26: Released v9.0.13